backend/pyproject.toml
[tool.poetry]
name = "django-skeleton"
version = "0.1.0"
description = "Django skeleton with most useful functionalities for new projects."
authors = [""]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.12.0"
Pillow = "^9.1.0"
sentry-sdk = "^1.5.12"
celery = "^5.2.6"
redis = "^4.3.1"
django-allauth = "^0.50.0"
dj-rest-auth = "^2.2.4"
django-extensions = "^3.1.5"
django-anymail = "^8.5"
django-phonenumber-field = "^6.1.0"
kombu = "^5.2.4"
raven = "^6.10.0"
django-filter = "^22.1.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.4.0"
drf-yasg = "^1.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"